refactor(Header) : emoji picker 컴포넌트 분리
